[LOG-12179] Add Gemfile.lock
Not 100% sure whether this should be included, since some people on
Stack mention that if you're developing a Gem you shouldn't check it in.

But without checking in, same as myself, the tests will probably wildly
fail for the next developer since all the dependendencies aren't
documented and upgraded to the latest version.